Now the claim export worker simply delegates to the ClaimsExportService
